// src/components/ai/types/contractTypes.ts
export interface SOPFormData {
  paymentClaimReferenceNumber: string;
  paymentClaimDate: string;
  respondentName: string;
  respondentAddress: string;
  respondentPhone: string;
  respondentFax: string;
  respondentEmail: string;
  respondentPersonInCharge: string;
  claimantName: string;
  claimantAddress: string;
  claimantPhone: string;
  claimantFax: string;
  claimantEmail: string;
  claimantPersonInCharge: string;
  projectTitle: string;
  contractIdentification: string;
  contractNumber: string;
  contractDate: string;
  referencePeriodFrom: string;
  referencePeriodTo: string;
  workCarriedOut: string;
  unfixedGoods: string;
  nominatedSubcontractor: string;
  unfixedGoodsNominated: string;
  totalAmount: string;
  lessAmountPaid: string;
  claimAmount: string;
  claimAmountWords: string;
}